+++ packages/lib/src/importer/mod.rs
@@ -5,8 +5,6 @@ mod require_waker;
 
 use crate::builtins::{self, top_level};
 
-const BUILTINS_AS_GLOBALS: &[&str] = &["fs", "net", "process", "stdio", "task"];
-
 pub fn create(lua: &'static Lua, args: Vec<String>) -> LuaResult<()> {
     // Create all builtins
     let builtins = vec![
@@ -19,14 +17,6 @@ pub fn create(lua: &'static Lua, args: Vec<String>) -> LuaResult<()> {
         #[cfg(feature = "roblox")]
         ("roblox", builtins::roblox::create(lua)?),
     ];
-
-    // TODO: Remove this when we have proper LSP support for custom
-    // require types and no longer need to have builtins as globals
-    let lua_globals = lua.globals();
-    for name in BUILTINS_AS_GLOBALS {
-        let builtin = builtins.iter().find(|(gname, _)| gname == name).unwrap();
-        lua_globals.set(*name, builtin.1.clone())?;
-    }
 
     // Create our importer (require) with builtins
     let require_fn = require::create(lua, builtins)?;
@@ -42,6 +32,7 @@ pub fn create(lua: &'static Lua, args: Vec<String>) -> LuaResult<()> {
     ];
 
     // Set top-level globals
+    let lua_globals = lua.globals();
     for (name, global) in globals {
         lua_globals.set(name, global)?;
     }